refactor: Separate FarePolicy
