+++ src/main/java/nextstep/subway/path/fare/policy/FarePolicies.java
@@ -11,36 +11,27 @@ import java.util.List;
 @Component
 public class FarePolicies {
 
-    private final List<FarePolicy> farePolicyList;
+    private final DistanceFarePolicies distanceFarePolicies;
+    private final LineFarePolicy lineFarePolicy;
 
-    public FarePolicies(List<FarePolicy> farePolicyList) {
-        this.farePolicyList = farePolicyList;
-    }
-
-    public Fare calculate(Distance distance) {
-        return farePolicyList.stream()
-                .map(farePolicy -> farePolicy.calculateFare(distance))
-                .reduce(Fare.ZERO, Fare::add);
-    }
-
-    public Fare calculate(int distance) {
-        return calculate(Distance.valueOf(distance));
+    public FarePolicies(DistanceFarePolicies distanceFarePolicies, LineFarePolicy lineFarePolicy) {
+        this.distanceFarePolicies = distanceFarePolicies;
+        this.lineFarePolicy = lineFarePolicy;
     }
 
     public Fare calculate(Path path) {
         Fare distanceFare = getDistanceFare(path);
         Fare lineFare = getLineFare(path);
-
         return distanceFare.add(lineFare);
     }
 
     private Fare getLineFare(Path path) {
         List<Station> stations = path.getStations();
-        return Fare.ZERO;
+        return lineFarePolicy.calculate(stations);
     }
 
     private Fare getDistanceFare(Path path) {
         Distance distance = path.getDistance();
-        return calculate(distance);
+        return distanceFarePolicies.calculate(distance);
     }
 }
